backend/user-service/src/services/user/user.service.ts
/**
 * @fileoverview TS 文件 user.service.ts 的功能描述
 * @author Team
 * @copyright 2024 组织名称
 * @license ISC
 */

@injectable()
export class UserService implements IUserService {
  constructor(
    @inject(TYPES.Logger) private logger: Logger,
    @inject(TYPES.Redis) private redis: RedisClient,
    @inject(TYPES.UserRepository) private userRepo: UserRepository,
  ) {}

  async findById(id: string): Promise<User> {
    // 实现用户查询逻辑
  }

  async updateProfile(id: string, data: UpdateProfileDTO): Promise<User> {
    // 实现资料更新逻辑
  }
}
